+++ waspc/data/Generator/templates/server/src/auth/providers/config/keycloak.ts
@@ -68,13 +68,11 @@ const _waspConfig: ProviderConfig = {
 
         return createOAuthProviderRouter({
             provider,
-            stateTypes: ['state', 'codeVerifier'],
+            oAuthType: 'OAuth2WithPKCE',
             userSignupFields: _waspUserSignupFields,
             getAuthorizationUrl: ({ state, codeVerifier }) => keycloak.createAuthorizationURL(state, codeVerifier, config),
-            getProviderInfo: async ({ code, codeVerifier }) => {
-                const { accessToken } = await keycloak.validateAuthorizationCode(code, codeVerifier);
-                return getKeycloakProfile(accessToken);
-            },
+            getProviderTokens: ({ code, codeVerifier }) => keycloak.validateAuthorizationCode(code, codeVerifier),
+            getProviderInfo: ({ accessToken }) => getKeycloakProfile(accessToken),
         });
     },
 }