+++ waspc/e2e-test/test-outputs/waspComplexTest-golden/waspComplexTest/.wasp/out/sdk/wasp/server/_types/taggedEntities.ts
@@ -7,17 +7,14 @@ import {
   type Entity, 
   type EntityName,
   type User,
-  type SocialLogin,
   type Task,
 } from 'wasp/entities'
 
 export type _User = WithName<User, "User">
-export type _SocialLogin = WithName<SocialLogin, "SocialLogin">
 export type _Task = WithName<Task, "Task">
 
 export type _Entity = 
   | _User
-  | _SocialLogin
   | _Task
   | never